requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.3.1
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
pymongo==4.5.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
six==1.16.0
yarl==1.9.2
